requirements.txt
Django==5.1.2
groq==0.11.0
pinecone==5.3.1
pinecone-client==5.0.1
python-dotenv==1.0.1
wikipedia==1.4.0

djangorestframework~=3.15.2
drf-yasg~=1.21.8